requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pymysql==1.0.3
bcrypt==4.0.1
